+++ cross/auto_parameters/__init__.py
@@ -1,21 +1,25 @@
 from .categorical_features import CategoricalEncodingParamCalculator
-from .clean_data import (
-    MissingValuesIndicatorParamCalculator,
-    MissingValuesParamCalculator,
+from .distribution_n_scale import (
+    NonLinearTransformationParamCalculator,
+    NormalizationParamCalculator,
+    QuantileTransformationParamCalculator,
+    ScaleTransformationParamCalculator,
 )
-from .datetime_features import DateTimeTransformerParamCalculator
 from .features_reduction import (
     ColumnSelectionParamCalculator,
     DimensionalityReductionParamCalculator,
 )
+from .missing_n_outliers import (
+    MissingValuesIndicatorParamCalculator,
+    MissingValuesParamCalculator,
+    OutliersParamCalculator,
+)
 from .numerical_features import (
     MathematicalOperationsParamCalculator,
-    NonLinearTransformationParamCalculator,
-    NormalizationParamCalculator,
     NumericalBinningParamCalculator,
-    OutliersParamCalculator,
-    QuantileTransformationParamCalculator,
-    ScaleTransformationParamCalculator,
     SplineTransformationParamCalculator,
 )
-from .periodic_features import CyclicalFeaturesTransformerParamCalculator
+from .periodic_features import (
+    CyclicalFeaturesTransformerParamCalculator,
+    DateTimeTransformerParamCalculator,
+)