+++ cross/auto_parameters/categorical_features/categorical_enconding.py
@@ -1,3 +1,5 @@
+from collections import defaultdict
+
 from tqdm import tqdm
 
 from cross.auto_parameters.shared import evaluate_model
@@ -8,54 +10,26 @@ from cross.transformations.utils.dtypes import categorical_columns
 
 class CategoricalEncodingParamCalculator:
     def calculate_best_params(
-        self, x, y, model, scoring, direction, cv, groups, verbose
+        self, X, y, model, scoring, direction, cv, groups, verbose
     ):
-        columns = categorical_columns(x)
-        encodings = [
-            "backward_diff",
-            "basen",
-            "binary",
-            "catboost",
-            "count",
-            "dummy",
-            "glmm",
-            "gray",
-            "hashing",
-            "helmert",
-            "james_stein",
-            "label",
-            "loo",
-            "m_estimate",
-            "onehot",
-            # "ordinal",
-            "polynomial",
-            "quantile",
-            "rankhot",
-            "sum",
-            "target",
-            "woe",
-        ]
-
         best_transformation_options = {}
+        cat_encodings = self._select_categorical_encodings(X)
 
-        with tqdm(total=len(columns) * len(encodings), disable=not verbose) as pbar:
-            for column in columns:
-                best_score = float("-inf") if direction == "maximize" else float("inf")
-                best_encoding = None
+        for column, encodings in tqdm(cat_encodings.items(), disable=not verbose):
+            best_score = float("-inf") if direction == "maximize" else float("inf")
+            best_encoding = None
 
-                for encoding in encodings:
-                    pbar.update(1)
+            for encoding in encodings:
+                transformation_options = {column: encoding}
+                handler = CategoricalEncoding(transformation_options)
+                score = evaluate_model(X, y, model, scoring, cv, groups, handler)
 
-                    transformation_options = {column: encoding}
-                    handler = CategoricalEncoding(transformation_options)
-                    score = evaluate_model(x, y, model, scoring, cv, groups, handler)
+                if is_score_improved(score, best_score, direction):
+                    best_score = score
+                    best_encoding = encoding
 
-                    if is_score_improved(score, best_score, direction):
-                        best_score = score
-                        best_encoding = encoding
-
-                if best_encoding:
-                    best_transformation_options[column] = best_encoding
+            if best_encoding:
+                best_transformation_options[column] = best_encoding
 
         if best_transformation_options:
             categorical_encoding = CategoricalEncoding(best_transformation_options)
@@ -65,3 +39,43 @@ class CategoricalEncodingParamCalculator:
             }
 
         return None
+
+    def _select_categorical_encodings(self, X):
+        cat_columns = categorical_columns(X)
+        category_counts = {col: X[col].nunique() for col in cat_columns}
+
+        selected_encodings = defaultdict(list)
+
+        for col, count in category_counts.items():
+            encodings = [
+                "basen",
+                "binary",
+                "catboost",
+                "count",
+                "glmm",
+                "gray",
+                "hashing",
+                "james_stein",
+                "label",
+                "loo",
+                "m_estimate",
+                "quantile",
+                "target",
+                "woe",
+            ]
+            if count <= 15:
+                encodings.extend(
+                    [
+                        "backward_diff",
+                        "dummy",
+                        "helmert",
+                        "onehot",
+                        "polynomial",
+                        "rankhot",
+                        "sum",
+                    ]
+                )
+
+            selected_encodings[col] = encodings
+
+        return selected_encodings